+++ apps/web/src/lib/api-client.ts
@@ -1,25 +1,9 @@
 "use server";
 
 import { cookies } from "next/headers";
+import { ApiError } from "./api-error";
 
 const API_BASE_URL = process.env.API_URL || "http://localhost:8000";
-
-/**
- * API 에러 클래스
- * status: HTTP 상태 코드
- * statusText: HTTP 상태 텍스트
- * body: 응답 본문 (JSON 파싱 가능한 경우)
- */
-export class ApiError extends Error {
-  constructor(
-    public readonly status: number,
-    public readonly statusText: string,
-    public readonly body?: unknown,
-  ) {
-    super(`API 요청 실패: ${status} ${statusText}`);
-    this.name = "ApiError";
-  }
-}
 
 async function handleErrorResponse(response: Response): Promise<never> {
   let body: unknown;
@@ -70,6 +54,7 @@ export async function apiClient(
   }
 
   return fetch(url, {
+    cache: "no-store",
     ...options,
     headers,
     credentials: "include",
